# api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: 'Avaliação Prática Lemobs'
  version: '1.0.0'
  title: Desafio API Lemobs

basePath: /v1/

consumes:
  - application/json
produces:
  - application/json

schemes:
  - http
  
parameters:
  usuario-id:
    name: usuario_id
    in: path
    description: id do usuario
    type: integer
    format: int32
    required: true

paths:
  # ---
  # Serviço responsável por cadastrar um novo usuário.
  # ---  
  /usuario:
    x-swagger-router-controller: usuario
    post:
      tags:
        - Usuario
      summary: Cadastrar
      operationId: cadastrar_usuario
      parameters:
        - in: body
          name: body
          description: Modelo de Usuário
          required: true
          schema:
            $ref: "#/definitions/usuario"        
      responses:
        '201':
          description: Criado (Created)
        '400':
          description: Requisição Inválida (Bad Request)
        '500':
          description: Erro Interno do Servidor (Internal Server Error)
    # ---
    # Serviço responsável por retornar dados de todos usuarios
    # ---
    get:
      tags:
        - Usuario
      summary: Obter todos os cadastros
      operationId: obter_todos
      responses:
        '201':
          description: Criado (Created)
          schema:
            $ref: "#/definitions/usuario"
        '400':
          description: Requisição Inválida (Bad Request)
        '500':
          description: Erro Interno do Servidor (Internal Server Error)
  # ---
  # Serviço responsável por atualizar um usuário.
  # ---
  /usuario/{usuario_id}:
    x-swagger-router-controller: usuario
    put:
      tags:
        - Usuario
      summary: Atualizar dados
      operationId: atualizar_usuario
      parameters:
        - $ref: "#/parameters/usuario-id"
        - name: "body"
          in: "body"
          description: "Updated user object"
          required: true
          schema:
            $ref: "#/definitions/atualizar"
      responses:
        '201':
          description: Atualizado (Created)
        '400':
          description: Requisição Inválida (Bad Request)
        '500':
          description: Erro Interno do Servidor (Internal Server Error)
    # ---
    # Serviço responsável por atualizar um usuário.
    # ---
    get:
      tags:
        - Usuario
      summary: Obter dados
      operationId: obter_usuario
      parameters:
        - $ref: "#/parameters/usuario-id"
      responses:
        '201':
          description: Sucesso ao retornar dados (Created)
          schema:
            $ref: "#/definitions/usuario"
        '400':
          description: Requisição Inválida (Bad Request)
        '500':
          description: Erro Interno do Servidor (Internal Server Error)
  
  /usuario/{usuario_id}/endereco:
    x-swagger-router-controller: usuario
    get:
      tags:
        - Endereço
      summary: Retorna endereços de um usuario
      operationId: obter_endereco
      parameters:
        - $ref: "#/parameters/usuario-id"
      responses:
        '201':
          description: Sucesso ao retornar dados (Created)
          schema:
            $ref: "#/definitions/enderecos"
        '400':
          description: Requisição Inválida (Bad Request)
        '500':
          description: Erro Interno do Servidor (Internal Server Error)
  
  # --
  # Serviço de cadastramento de endereço
  # --
  /endereco:
    x-swagger-router-controller: endereco
    post:
      tags: 
        - Endereço
      summary: Cadastrar endereço
      operationId: cadastrar_endereco
      parameters:
        - in: body
          name: endereco
          description: Modelo de Endereco
          required: true
          schema:
            $ref: "#/definitions/endereco"        
      responses:
        '201':
          description: Criado (Created)
        '400':
          description: Requisição Inválida (Bad Request)
        '500':
          description: Erro Interno do Servidor (Internal Server Error)
    
    # --
    # Serviço responsavel por retornar todos os endereços
    # --
    get:
      tags:
        - Endereço
      summary: Retorna dados de todos endereços
      operationId: obter_enderecos
      responses:
        '201':
          description: Feito!
          schema:
            $ref: "#/definitions/endereco"
        '400':
          description: Requisição Inválida (Bad Request)
        '500':
          description: Erro Interno do Servidor (Internal Server Error)
      
  
  /docs:
    x-swagger-pipe: swagger_raw
    
definitions:
  usuario:
    type: object
    properties:
      nome:
        type: string
      data_nascimento:
        type: string
        format: date
      cpf:
        type: string
    required:
      - nome
      - data_nascimento
      - cpf
  atualizar:
    type: object
    properties:
      nome:
        type: string
      data_nascimento:
        type: string
        format: date   
    required:
      - nome
      - data_nascimento
  endereco:
    type: object
    properties:
      rua: 
        type: string
      numero: 
        type: integer
        format: int64
      complemento:
        type: string
      bairro:
        type: string
      usuario_id:
        type: integer
        format: int64
    required:
      - rua
      - bairro
      - usuario_id
  enderecos:
    type: object
    properties:
      total:
        type: integer
        format: int32
      enderecos:
        type: array
        items:
          type: object
          properties:
            endereco:
              type: string
            bairro:
              type: string